+++ metadoc/clojure2d/pixels_example.clj
@@ -367,7 +367,7 @@
 (def local-noise (r/fbm-noise {:seed 42}))
 
 (add-examples renderer
-  (example-snippet "Usage" saver :image
+  (example-snippet "Log density rendering" saver :image
     #(let [r (renderer 300 300)]
        (dotimes [i 20000000]
          (let [x (+ 150 (r/grand 30))
@@ -376,6 +376,24 @@
                y (+ y (* 20.0 (- ^double (local-noise (/ y 50.0) (/ x 50.0) 2.23) 0.5)))]
            (add-pixel! r x y :white)))
        (to-pixels r {:gamma-alpha 0.6 :gamma-color 0.8})))
+  (example-snippet "Linear density rendering" saver :image
+    #(let [r (renderer 300 300)]
+       (dotimes [i 20000000]
+         (let [x (+ 150 (r/grand 30))
+               y (+ 150 (r/grand 30))
+               x (+ x (* 20.0 (- ^double (local-noise (/ x 50.0) (/ y 50.0) 0.34) 0.5)))
+               y (+ y (* 20.0 (- ^double (local-noise (/ y 50.0) (/ x 50.0) 2.23) 0.5)))]
+           (add-pixel! r x y :white)))
+       (to-pixels r {:gamma-alpha 0.6 :gamma-color 0.8 :linear? true})))
+  (example-snippet "Splat rendering" saver :image
+    #(let [r (renderer 300 300)]
+       (dotimes [i 20000000]
+         (let [x (+ 150 (r/grand 30))
+               y (+ 150 (r/grand 30))
+               x (+ x (* 20.0 (- ^double (local-noise (/ x 50.0) (/ y 50.0) 0.34) 0.5)))
+               y (+ y (* 20.0 (- ^double (local-noise (/ y 50.0) (/ x 50.0) 2.23) 0.5)))]
+           (add-pixel! r x y :white)))
+       (to-pixels r {:gamma-color 0.8 :splats? true})))
   (example-snippet "Compare to native Java2d rendering. You can observe oversaturation." saver :image
     #(let [r (core/black-canvas 300 300)]
        (core/with-canvas [c r]